BASE-17: creating standard api routes and refactor web routes
