+++ resources/views/admin/groups/index.blade.php
@@ -3,7 +3,7 @@
 @section('content')
     <div class="p-6">
         <div class="relative overflow-x-auto shadow-md sm:rounded-lg">
-            <table id="table_id" class="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+            <table id="table_id2" class="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
                 <thead class="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
                     <tr>
                         <th scope="col" class="px-6 py-3">
